召唤兽技能，治疗技能的完善

+++ gameserver/src/main/java/com/wan37/gameserver/event/model/AttackMonsterEvent.java
@@ -23,7 +23,7 @@ public class AttackMonsterEvent extends Event {
     private Player player;
     private Monster target;
     private GameScene gameScene;
-    private Integer damage;
+    private Long damage;
 
 
 }
